Revert "renamed rest api"
